Started hard_cpu w/ minmax (needs a lot of work)
